Performance Commit 2

+++ src/main/java/WayofTime/alchemicalWizardry/common/spell/complex/effect/impactEffects/earth/ProjectileDefensiveEarth.java
@@ -22,8 +22,8 @@ public class ProjectileDefensiveEarth extends ProjectileImpactEffect
     @Override
     public void onTileImpact(World world, MovingObjectPosition mop)
     {
-        int horizRange = (int) (this.powerUpgrades);
-        int vertRange = (int) (this.potencyUpgrades);
+        int horizRange = this.powerUpgrades;
+        int vertRange = this.potencyUpgrades;
 
         int posX = mop.blockX;
         int posY = mop.blockY;
